bodypix_requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
pillow>=9.0.0
opencv-python>=4.8.0
watchdog>=3.0.0
tf-bodypix>=0.1.0
